+++ lib/generators/templates/add_pilipinas_code_indexes.rb
@@ -1,8 +1,8 @@
-class AddPilipinasCodeIndexes < ActiveRecord::Migration<%= migration_version %>
+class AddPilipinasLocationIdIndexes < ActiveRecord::Migration<%= migration_version %>
   def change
-    add_index :pilipinas_regions,   :code, unique: true, if_not_exists: true
-    add_index :pilipinas_provinces, :code, unique: true, if_not_exists: true
-    add_index :pilipinas_cities,    :code, unique: true, if_not_exists: true
-    add_index :pilipinas_barangays, :code, unique: true, if_not_exists: true
+    add_index :pilipinas_regions,   :location_id, unique: true, if_not_exists: true
+    add_index :pilipinas_provinces, :location_id, unique: true, if_not_exists: true
+    add_index :pilipinas_cities,    :location_id, unique: true, if_not_exists: true
+    add_index :pilipinas_barangays, :location_id, unique: true, if_not_exists: true
   end
 end